node: refactor with_reader -> as_reader, etc.
Doing conditional logic like this with functions that take closures that
they only conditionally call isn't really idiomatic rust, and comes with
it a lot of issues such as not always working well with result
propagation, not meshing with async, etc. - generally, the thing to do
is return an optional borrowed reference which the caller can then match
on and do stuff with, and let borrowck figure out all the lifetimes of
the references (something that especially with NLL it's very good at!).

In the interest of making things easier to work with, moving to more
idiomatic Rust usage, and removing some (unnecessary) panics+unwraps,
this commit refactors the `with_reader`, `with_reader_mut`, etc. methods
on Node into `as_reader`, `as_mut_reader`, etc. respectively, which
return optional references instead of taking a closure.