Document and allow panics for State invariants
Because it's all highly optimized code, the various internals of the
various kinds of State include several instances of explicit panics in
cases where the code path is truly unreachable. This commit explicitly
allows all of those panics, and adds a bunch of extra documentation for
what those invariants *are*, both internal and external.