Filter out columns not requested in SELECT
This change filters out columns that are projected in order to execute a
query but not requested in the query itself. The set of columns to be
returned to the client is built from the parsed nom_sql select
statement. The columns are stored in the PostLookup struct and used for
filtering in PostLookup::process(). The set of columns are also passed
to the view to update the schema.

The ViewSchema now stores the set of returned columns and projected
columns. A SchemaType enum is introduced to select between the two
schemas.

PostLookup::process now returns a Vec<Vec<&DataTypes>> so that it can
manipulate the structure of the rows (i.e., remove columns).

Column prefixes that were previously added to filter out columns at the
client that were projected for order by clauses but not part of the
result are removed now that filtering is done at the reader.

Bogokey columns from integration tests have been removed because those
are no longer returned on lookup.