Setup framework pipeline for fallback
This commit sets up the framework pipeline for fallback. This includes
expanding the docker-compose file for readyset to deploy mysql as well.
Mysql wants to occupy the internal 3306 port, so rather than fight it,
in the noria case we let it occupy 3306 and set noria to 3333.

This means that for the simple mysql8 testing we also need to update it
to run on port 3333. No matter what it will also try to grab 3306, but
it shouldn't matter with no collision potentials.

This way we can update our "app" (each framework being tested) to point
at 3333, and any end "db" (noria or mysql) can be communicated over said
port.

I had to also change the ENV VAR listing to the recommended format for
docker compose files. I discovered that the way we had been listing them
didn't allow for full fledged mysql endpoint urls to get interpreted in
whole.