Run desugaring passes prior to subquery flattening
Previously, the detect_problematic_self_joins pass was run as part of
the rest of the desugaring passes, *after* queries had their subqueries
flattened into references to views in the query graph, meaning it was
basically incapable of detecting problematic self joins, since almost
all of them were happening through subqueries. Moving it above the
subquery flattening allows it to detect those cases, but also requires
moving all of the desugaring passes above subquery flattening as well,
since detect_problematic_self_joins can only run on a desugared
query (it needs expand_implied_tables and expand_stars at the very
least). That in turn required updating both expand_implied_tables and
expand_stars to operate on subqueries, both directly in joins and in
CTEs. To aid in testing this change, this commit refactors the testing
in the star_expansion pass to be a little bit easier to add new test
cases too.

With this change, we now properly detect cases of problematic self
joins in subqueries issued directly as part of a recipe.