Setup js-client testing for fallback
This commit sets up js-client testing for fallback, by ensuring that we
are passing down the replication url, and deploying a mysql instance
correctly for fallback.

This also adds a healthcheck for noria-server, and chains health check
dependencies. This has the same effect as making sure js-client waits
for zookeeper, mysql, and noria-server. It simply ensures those waits
are contingent on health checks coming back healthy.