Document future refactor of packet types
This commit allows unreachables for the time being in Node process()
function and documents each unreachable. These unreachables are the
result of matching on the Packet enum, and finding we have been given a
packet type that should have never been passed to this function.

Ideally we should have an enum of enums so each variant is a valid type
and we can leverage the type system rather than have to many
unreachables.

That's likely a huge refactor though, and has been documented already
in: https://readysettech.atlassian.net/browse/ENG-455