Document on_connected invariant on many methods
This commit documents an invariant that must be enforced on these three
helper methods to not get panics.

The call sites for these were so extensive that it made more sense to do
this than pollute the codebase with very excessive unwraps (and allow
unwraps with comments on each).

In all places in the codebase currently calling these, on_connected has
already been called, so these are safe.