Refactor Deref and DerefMut on Node
This commit removes the Deref and DerefMut implementations that were on
Node. These had internal panics, when Deref and DerefMut should never
panic.

Instead I created as_internal() and as_mut_internal() helper methods,
because what the Deref and DerefMut were really doing, is matching on
self, and only if internal, passing back the inner operator. This does
the same thing, but in cases those implementations would have panicked,
we instead pass back None and force the caller to deal with it.

This commit is extensive because this affects many call sites. I tried
at most call sites to adapt the function to return a ReadySetResult, and
pass back new and appropriate errors. In a couple places I was hesitant
to change the return signature for functions that were clearly commands
(in line with the Command-Query separation principle).