Create a base Docker Compose configuration
In order to support working on working on alternative consensus systems,
a development environment is needed for working on that.

Consul suggests using Docker Compose and we are already using it for the
build process. This extracts out the services needed to run all the
tests into a separate `docker-compose.yml` which now
`build/docker-compose.ci-test.yaml` now overrides to support building
and running the tests inside of the container.

Initial configuration has been created for running a Consul server but
this is still in progress.