Make Domain::resume_polling a method
There's a TODO in worker/replica.rs in the bit that passes ResumePolling
to Domain::on_event that suggests that refactoring it to be a method
rather than just an enum of PollEvent would be a good idea - and it
turns out it was! Doing this allows us to avoid an `unreachable` when
matching on the return type, simplify the code at the callsite
significantly, and simplify the signature of the resume_polling method
itself!