Add a timeout of 6 hours to the fuzz pipeline
Occasionally this pipeline finds a bug that causes it to get stuck,
sometimes for multiple days - clogging up the agent pool and wasting a
bunch of cloud money. 6 hours feels like more than enough to find actual
bugs, but not so much that we're just pouring compute time down the
drain.