Refactor noria-server to use Clap v3 derive macro
Refactor the argument parsing in the main function of the noria-server
binary to use the clap v3 beta "structopt-style" derive macro, instead
of the older builder-style definition. This gets us a lot of readability
and maintainability, but more importantly allows us to extract and parse
configuration with basically no further work, and without using unwraps,
panics, or unreachables at all.