Remove all code related to security universes
Remove everything related to security universes, including:

- security policies on recipes
- the creation of and handling of security universes in mir
- table mappings throughout mir
- the Rewrite operator

Security universes aren't a feature that we're shipping in the current
iteration of the system, nor are we planning on doing so for the
forseeable future, and having them present in the codebase is creating
a maintenance burden that we'd like to avoid - specifically in this case
the code for dealing with security universes had a fair number of
panics, which we're currently working towards removing from the codebase
entirely. Deleting security universes removes all of those panics,
significantly simplifies a large portion of the codebase and, since they
were seemingly completely untested, leaves everything working as-is.