Grouped nodes can only have a single over column
Grouped nodes, including group concat, are only ever constructed with a
single over column, and there's an assertion in schema that they only
ever *return* one over column, so having over_columns return a vec (and
storing a vec of over columns internally to group_concat) is rather
silly. Refactoring them to only have a single over column removes the
necessity for the assertion in schema, plus a couple of usages of
indexing_slicing.