Avoid panics in schema.rs
Remove all potential panics in controller/schema.rs. Mostly, this was
done by replacing `unreachable!()`s in `type_for_internal_column` with
just tracing through to the next node in the graph - these *may* have
actually been unreachable, but in the case that we actually hit them
it's reasonable to just keep tracing through to find more columns. Also,
the error for if the schema for a base table is a View has been turned
into an actual error, though if this really happens in production I'm
not sure what we're capable of doing.