Make extract_subqueries pass fallible
The extract_subqueries pass needs to be fallible so it can return an
`unsupported` error in case it comes upon an EXISTS expression. Once we
implement EXISTS fully we can revert this, but for now we need to be
able to return an error for queries like this so we can fallback to the
underlying db.