Avoid panicking when creating new base tables
It's (obviously) only valid to pass a CREATE TABLE statement to the
`named_base_to_mir` and `add_base_via_mir` functions - so having these
take SqlQuery instead of CreateTableStatement then panic if it's
anything else is a little silly. Converting the argument type to
CreateTableStatement avoids a panic *and* an error.