Use LogicalOp in classify_conditionals
Pull the LogicalOp enum that we had defined to avoid an unreachable!()
call after pattern-matching on only logical operators up to query_utils,
so it can be used in classify_conditionals as well. The same general
idea applies here - rather than checking that a BinaryOperator is one of
And or Or and then panicking with `unreachable()` later in a pattern
match, we *convert* it to a more restrictive enum first and then take
advantage of rustc's pattern match totality checking to avoid the
`unreachable!()` entirely.