Convert unreachable to debug_assert for improbable
This converts an unreachable into a debug_assert case. The struct itself
has private internal fields, and the impl declaration when followed
shows that this really is unreachable.

We can't have > 0 memory and not have anything stored in our map, which
is evidenced by following when memory gets stored.

This should save us a needless computation and provide the functionality
the initial author likely wanted - which was a sanity check during
development.