Return error for real potential bug
This is a scenario we could actually hit in production, and we want to
be able to deal with the error explicitely in the future when we know
how to deal with this particular edge case.

Mark_filled clears the hole, so if this happens it means we got two
replays for the same key, which could hypothetically contain different
data.

There's no way for us to know at this point in scope which is the
correct version. Something has gone horribly wrong, and if we ever hit
this specific error, we really need to dig into why we got here.