Update rocksdb version
The current version (0.14) is using RocksDB 6.7.3, this one (0.16)
is using 6.17.3. There has be many performance improvements inbetween
especially to improve p99.99.

In addition 0.16 adds an option to read multiple kyes in one batch,
something that we should incorporate in the future.

Latest version of this crate is 0.17 but upgrading to it is causing
conflicts between funty and bitvec :(