Delete from record cache when topk receives deletes
If a fully materialized topk (which is caching records past k so it can
backfill if deletes ever bring it below K records) receives deletes that
*don't* match any of the records in the top K, it needs to apply those
deletes to the *cache* itself, otherwise when subsequent deletes do end
up requiring a backfill we'll backfill with records that have been
deleted.

Fixing this fixes 3 previously-failing generated logictests.