Extended the read/write experiments to report metrics to InfluxDB.
We now have the possibility to run the experiments and report metrics to an
InfluxDB instance. This is optional, and if no InfluxDB configuration is
provided, no metrics are reported.
Another change was the addition of a `cycles` option, allowing to configure how
many times the experiment should be run (as oposed to running forever as it is
now). This is also optional, and if no configuration is provided, the
experiments run until interrupted (just as before).

Change-Id: I1f2b64ae5c561ed58658c134ae69655e6aaca599
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/138
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>