Check readers too for full-below-partial case
It's possible, though unlikely, that as part of a migration we end up
creating a fully materialized reader that's a desccendant of a partially
materialized node but doesn't have any other new materialized nodes in
between it and its partial ancestor - in my testing, I've seen this
happen specifically when a fully materialized reader is added below a
partially materialized *ingress*, but it's possible there are other
conditions that trigger it. If that does happen, previously the bit in
commit() that checked for partially materialized nodes above fully
materialized nodes wouldn't be able to detect it, because it was only
iterating over `self.added` to find new stateful nodes, and since
reader state is special readers don't get added to `self.added`. Now, as
part of figuring out indices for readers, we also track newly added and
materialized readers in a new `self.new_readers` field, and iterate over
that in addition to `self.added` when checking for full-below-partial.
This causes the error when this happens to occur during the migration
rather than when we attempt to initiate a full replay starting at a
partially materialized node, leaving the domain running happily rather
than fallen over on its face.

Unfortunately, similar to the rest of the code with migrations, this is
really difficult to test automatically, but I've tested it manually with
the following logictest and it does the correct thing:

    statement ok
    CREATE TABLE table_3 (
        column_2 INT(32),
        column_1 INT(32),
        column_3 INT(32),
        PRIMARY KEY (column_3)
    )

    statement ok
    CREATE TABLE table_1 (
        column_3 INT(32),
        column_2 INT(32),
        column_1 INT(32),
        column_4 INT(32),
        PRIMARY KEY (column_4)
    )

    statement ok
    SELECT
        table_1.column_3 AS alias_1,
        table_3.column_2 AS alias_2
    FROM table_1
    INNER JOIN table_3
    ON (table_1.column_2 = table_3.column_1)
    WHERE (table_1.column_3 = ?)

    statement ok
    SELECT
        table_1.column_3 AS alias_1,
        table_3.column_2 AS alias_2
    FROM table_1
    INNER JOIN table_3
    ON (table_1.column_2 = table_3.column_1)
    WHERE (table_1.column_3 = 1)