noria/server:  Handle single-aggregation queries when there are no rows present
This is implemented in the reader.  We make each aggregation type (e.g.
COUNT(), SUM(), etc) aware of its "default" value; add a list of aggregations to
PostLookup; and populate that list of aggregations in
mir_node_to_flow_parts(), where we're currently creating the PostLookup.

While processing the PostLookup, if there are both zero rows AND at
least one aggregation, we short-circuit and just return all the
aggregations' default values.