Remove Epoch and improve LocalAuthority
This is the first PR in the process of supporting Consul and generally
improving how we manage Noria workers and controllers in general.

The core part of this is to make LocalAuthority into two parts to make
it "look" more like other Authority types that will be created. This
means LocalAuthority now has a state inside of a Mutex which will be
similar for other Authorities going forward. ZookeeperAuthority also
has state now as well.

Epoch has been removed as it was used for worker registration and
heartbeat, both of which will be managed by the Authority as this
process continues.

The next phase of this will be to move worker registration and listing
into being functions on Authority instead of stored by the Controller.

Change-Id: I5d1eb269cc35cf40b687acb1e5e147101bdef0a1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/158
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI