Convert target ignore to subdirectory catchall
It appears that simply haveing `target/` in our `.dockerignore` file
does not ignore target directories found inside individual workspaces.
This catch all ensures that we ignore all target directories regardless
of where they live in the mono-repo.

This helps reduce the amount we send to the docker daemon. It's quite
common to have nested target dirs if you are running cargo check from
inside a specific cargo project in the mono repo.