Pass back actual MySQL error
We should be passing back the actual MySQL error we get in the case of
fallback. This ensures that by making sure that we pull the ErrorKind
from the actual underlying MySQL error.

This also fixes an issue in `on_execute()` where we were returning any
errors we got, which caused dropped connections up the chain. We should
instead write the error back as a MySQL error to the client, and return
Ok() so we keep the connection open.