Add server's volume_id to the server and controller state.
The volume associated with a server contains persistent state
needed for recovery from failures. Adding the server's volume
id will allow future changes to use the volume associated with
newly started instances to recover from failures.

This adds the new server parameter `--volume_id`, and propagates
it in worker registration to the controller.

Change-Id: I8c5a2543c58179b9c064bbecc5deda7c9f49e86a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/61
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>