Cost savings and improvements for CI for Terraform
Most importantly is that it shifts the default for RDS instances back to
using General Purpose Storage instead of IO Provisioned Storage to save
money in general.

In testing for CI, I discovered that we could not create with many
names for environment and deployment since certain names are limited to
32 characters. Instead of using the name field, that will be random and
instead the same value will be associated with a Name tag.