Create temporary environment Terraform
This moves the readyset module into a modules directory and creates an
environments directory under ops/tf where instantiations of Terraform
modules can live.

This also fixes a couple of other bugs I found while testing temporary
environment creation.

The create-tmp-env.sh script is there mostly for educationional purposes
on how to use the code here for planning, applying, and destroying a
temporary environment.

The change around ASG tags is due to a bug which we hit at
hashicorp/terraform-provider-aws#14085