Expose the failing script when fuzz testing fails
When noria-logictest running in fuzz mode finds a failing test script,
save that script to a new temporary file. Then, if we're running in
buildkite (which we determine by checking if an env var is set), upload
that script to a new buildkite artifact by shelling out to the
`buildkite-agent artifact upload` command.

Actually getting at the generated test script required slightly
shuffling around the way fuzz testing is run - now, we have a proptest
`Strategy` that generates the `TestScript` *itself*, and that is used as
the input (and eventually counterexample) to the test property itself.
This means that if test script generation fails for whatever reason we
don't get to take advantage of scripting (unfortunately) because failing
strategies don't shrink, but it's imo a reasonable tradeoff for getting
us access to the failing test script.

Change-Id: I3b15e9af5790645ac00a2054d6cb5b587e08bc99