Delegate Display to underlying MySQL error
The way that our error macro crate was concatenating errors made for
really gross error messages to be written back to the MySQL client. This
delegates responsiblity for display formatting to the underlying from
error for cleaner error messages.

Separately it appears as though our mysql error writer method is writing
the mysql error code and state to the message itself. This means that we
actually need to match on underlying MySql error codes and explicitely
grab out the code and message so we don't get duplications in the
message.