Implement prepare/execute over mysql_connector
Prepare and execute statements are implemented for mysql_connector. The
mysql connector was updated to use a single connection rather than a
connection pool, because the connection to the mysql database stores the
state of all prepared queries.

A prepared statement id table is created in the backend struct to avoid
conflicts in noria and mysql prepared statement ids. The mysql_connector
stores a table mapping the statement id to mysql_async::statement for
execution. Since the backend stores a seperate reader/writer
mysql_connector, the query must go to the correct mysql_connector
instance.

The current method to determine whether a query executed on the
underlying mysql database will produce a read or writer result will not
correctly identify all read statements. A solution to identify the type
of result based on the mysql server response is in progress. This will
remove the need to have separate read/write functions in the
mysql_connector.

race_read() now builds a new mysql_connection on every call. This is
because the mysql_connector can not be cloned now that it implements a
single mysql connection (unclonable) rather than a connection pool.

DataType to mysql::Value conversion is implemented to pass params
through mysql_async. Implementations for mysql_async::Value and
mysql::Value were added, but the implementation for mysql_async should
be removed once all mysql_common exports are resolved to the same
version.

Finally, this commit also removes an expected-fail logictest that as of
this commit stopped failing with fallback enabled - in the future, we'll
have the ability to conditionally skip or not-skip tests based on
whether we're running with fallback enabled, but for now all tests run
both with and without fallback so have to either always fail or always
pass regardless of whether fallback is enabled.

Change-Id: I16b8714f8a4559156438c4082eea099ce4444672
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/155
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI