Domain stickiness when `volume_id` parameter is passed.
This changes domain assignment to be "sticky" for base tables,
that is, domains with base tables will always be assigned to
the same volume id after initial assignment in a deployment.
If this is not possible, for instance, if a server with the correct
volume id does not exist due to a failure, the migration adding
the base table node will throw an error.

This is accomplished by adding a new per-node,
`DomianPlacementRestriction` struct that is populated when base
tables are inserted into the dataflow graph during a migration.
This `DomainPlacementRestriction` struct is persisted in the
controller state and used during domain assignment to ensure that
base table nodes are placed into domains that can satisfy their
restrictions, and these domains are either on servers, or placed
into server that can meet their restrictions.

This currently will fail migrations that would previously have
left the system in an unrecoverable state. In a future change,
we will want to have the system re-attempt migration when this
occurs.

Change-Id: Ic125afb67e319dec779ce4ae884e53a75e929fcf
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/62
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>