Only reference tables in the query
Relying on `state.some_table_mut` to get tables to reference in a
query (eg for their columns, to filter on or project) breaks down in the
presence of other queries in a script or subqueries in the current
query. This commit adds a new `some_table_in_query_mut` method, which
only ever returns a table referenced in the passed-in query, and uses
this in all of the add_to_query implementations where relevant.

Change-Id: Iaa23e919f00344a3dea4f9d23752a329c0f1ef74