Add buildkite step to set Verified in gerrit
Add a script, configured as the final step in our buildkite pipeline,
that will use the gerrit REST API to review the CL+patchset for the
current build, if any, setting the Verified flag to either +1 or -1
depending on the status.

This uses a reasonably convoluted method to actually *get* the status -
buildkite knows if the steps before us have failed, but doesn't expose a
way of actually *getting* that information in a running step, so
we (ab)use the build "meta-data" to set a flag at the start of the build
that the pipeline has failed, and then in a post-command hook if the
command has failed toggle that to failed - then use that in the script
to determine the overall build status.

Change-Id: I6792d440fcbd52be60d8e499c8aef0e602e0c845
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/44
Reviewed-by: Harley Klein <harley@readyset.io>
Tested-by: Buildkite CI