Reimplement Replica with async code
Currently the custom event loop is quite complex and operates
on timeouts, this simplifies the event loop with async code
and handles events eagerly.

We still try to aggregate the packets that matter (RequestReaderReplay
only right now) into a single request, and other batching is done more
efficiently in the previous CL.

This big rewrite was triggered mainly by poor performance in the fastly
demo.

The fastly benchmark of this CL chain compared to the main branch ATM
shows that with 75% materialization read QPS improved from 3213 to 5408
with uniform distribution and from 47189 to 68360 with zipf distribution
which is a 40.5% and 31% improvement respectively.

Also before the commit the flamegraphs showed a lot of time spent spinning,
whereas now 90% of the time is spent in pure packet handling.

There are some very big changes here, but I feel pretty confident about them:

* `group_commit` was removed, as Input packets grouping should be done
  (and is done?) in the replicator.
* ack code for Input packets was moved to the main loop
* using `StreamMap` instead of `StreamUnordered`, which also allows us to use
  a 64bit token instead of a 32bit token with epoch.
* Timeout based spinning removed, except for the frontier purges.

Change-Id: I0eca133ef9306fb1f424f706e740fcb534eaeebc
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/184
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>