Removed silly check on `PersistenceParameters` fields.
This commit removes the assertion that makes the `PersistenceParameters`
creation to fail if there is a hyphen (`-`) in the `log_prefix` parameter (now
renamed to `db_filename_prefix` for clarity), but adds a transformation to
replace the hyphens for underscores (`_`) in the domain logic (which is the
place where the String needs to be sanitized, in order to create an identifier).
The reason is that we use Noria deployment name as the `db_filename_prefix` in
the `PersistenceParameters`, and now that we are automating deployments (by
using Terraform), we use hyphens in the name to have better deployment
identifiers (i.e., a deployment might be named following a
`[COMPANY]-[COMMIT]-[ENVIRONMENT]`).

Change-Id: I4a740502af242d58b271338ae9eb30e47438378d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/41
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>