Add gerrit ref-updated hook
Add the script that is currently being run as the gerrit ref-updated
hook, which uses the buildkite API to trigger a new buildkite build of
new or updated CLs, and newly merged CLs. This is here really just as
documentation - there's nothing that deploys from this location yet -
but it's better to have it in version control than to just have it on a
server.

Change-Id: I23e901377aa1ed3bd4fdc48d717015d16e53771f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/86
Reviewed-by: Harley Klein <harley@readyset.io>
Tested-by: Buildkite CI