Don't generate SqlType::Decimal, for now
The DECIMAL MySQL type is... a little odd, in its semantics; it's a
fixed-point floating type with a configurable fixed number of digits in
the type, including a configurable fixed number of digits after the
decimal point. Because of this, it's reasonably difficult to write a
correct proptest strategy to generate it. This just disables generating
the type, for now, though the two fields (which MySQL's docs helpfully
refer to as "m" and "n") do have strategies that at least make
the *type* generate correctly for if we want to re-enable it in the
future.

Change-Id: Ibb567f745b7fd1e4ff556825ad8e2d353a79aab4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/95
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>