query-generator: Add a test against mysql
Add a new integration test for query_generator, a proptest that
generates random queries using the Arbitrary impl for QuerySeed and
checks that all queries will at least execute against mysql without
returning an error.

This test is ignored as it currently fails due to a bug with timestamp
values that seems very difficult to diagnose, but I'm committing it
because it still provides some utility in trying to find issues where we
generate invalid queries.

Change-Id: Id71270c24595e9ccc9c1973142c183155a169f41
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/96
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>