Move insert and remove impl into KeyedState
Previously, the implementation of KeyedState and that of SingleState
were pretty much completely tightly coupled - you couldn't really ever
use a KeyedState that wasn't part of a larger KeyedState. I need that to
change, though, in order to support the addition of weak indices (which
will be backed by a bare KeyedState). To make KeyedState useful on its
own, this moves the bulk of the implementation of `insert` and
`remove_row` into methods on KeyedState, beefs up the documentation on
both, and cleans them up slightly. In both cases we pass in the row in
question plus the set of columns in the row that the key is being
derived from.

Change-Id: Icca9ad0d152e5ad2549e0964a86e3d0358d356ee
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/121
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>