Add weak keys to State
Add methods to the State trait to add and lookup into weak key indices,
and fully implement weak keys for both PersistentState and MemoryState.
These are currently unused in the rest of the code, however.

There's more information in the design doc linked in the source, but in
brief these are being added to support a bug where joins with partial
parents can fail to find rows to join on in one of those parents due to
a mismatch in indices, even when there are filled holes in downstream
state that we need to target writes to.

Change-Id: I395fee2392f1a3c6ef63f4822ba23d173e877815
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/122
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>