Allow Ingredients to request weak indices
Extend the `suggest_indexes` method on Ingredient to return a new
`SuggestedIndex` enum instead of just an Index, which has variants for
both Strict and Weak indices, and thread that all the way through both
materialization and the initialization of the indices of a State in
Domain to create weak indices in state when requested.

There's a couple of comments in the code as to this effect, but just to
belabor the point - when a node requests a weak index we *also* create a
strict index (in materialization, right now) with the same index type
and columns, since when processing replays we have to be doing lookups
into strict indices and not weak indices (since replays have to
represent all state for a given key, not just those rows that won't hit
holes downstream).

Change-Id: Iec571acc0532ca9aea3520dedf008843b22f7288
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/123
Reviewed-by: Michael Cronce <mc@readyset.io>
Tested-by: Buildkite CI