Use weak indices when processing joins
Add a new `mode: LookupMode` parameter to both Ingredient::lookup and
::query_through, specifying whether to look up into weak or strict
indices, and then use this when processing normal, non-replay deltas in
joins to use weak indices to perform the lookup into the *other* parent
to perform the join itself. This fixes a bug where, if a replay for a
key *other* than the join key has been performed *through* one parent of
a join, subsequent deltas from the *other* parent aren't able to find
any rows to join on and are incorrectly discarded. This commit adds a
new integration test that was failing due to that bug, and also renames
8 previously-failing logictests to no longer failing that were fixed by
this change.

Change-Id: I3723e5220ed241428d8f808e59d9c6ecd9c1d536
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/124
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>