Implement LIKE in dataflow expressions
Implement constant-valued LIKE, ILIKE, NOT LIKE, and NOT ILIKE in
dataflow expressions, all by calling out to the LikePattern implemented
in noria::util::like.

This is going to be slower than it otherwise could be because we have to
re-build the LikePattern on every evaluation rather than caching
it, which we could do if the right-hand side is a constant value rather
than a column reference - we *may* want to optimize that at some point
in the future (ENG-492 for that), but that's left as a point of future
improvement for now.

Change-Id: I09219319da5932927778eac615fccd881dbf1e17
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/135
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>