Initial Buildkite configuration with Rust runner
This eventually will be supporting running through Buildkite the many
combinatorial Terraform root modules that are created by Substrate.

I initially attempted to implement this using Python but Rust gave me a
slightly more secure footing to work from as I work on this project. The
Terraform Techniques with Buildkite blog post on the Buildkite blog used
rake and the closest thing in Python is Fabric which is not great.
Makefiles are also not a thing I enjoy and Substrate generates its own
already but are not exactly in a CI-able state.

This initial commit sets up a pipeline for the Rust code and then uses
it to generate a small pipeline from there which verifies that things
will work with this setup.