Refactor noria-adapter to use clap v3 derive macro
Refactor the noria-adapter crate's (and, by extension, both noria-psql's
and noria-mysql's which depend on it) command line argument parsing to
use the clap v3 derive macro rather than the older v2 builder pattern.
This is mostly just a lot more readable / maintainable, and more
consistent with the pattern we're using in the rest of the application.

In this refactor, both trace and verbose were dropped as supported
flags, since both turned out to be completely unused - besides that,
everything else is copied over basically verbatim

Change-Id: I5ce31d6b66ec6dfc267477f3ee447655dbc08cc2