Move serialization into Authority
Instead of Authority returning `Vec<u8>`s which then need to be
deserialized by the user of this abstraction, this moves the
serialization interally. This also allows us to in the future optimize
the serialization protocol and such through the abstraction.

There was a consideration of making the types to serialize to be
associated types but it would require massively refactoring a huge
amount of the code that uses Authority and inside of Noria there is
only a single type that would be made. If this eventually gets
abstracted into another crate, associated types would make sense there.

A raw read function was left as there is an endpoint that returns it.
This should be updated to better show what we believe the Authority to
know instead of just passing along queries to Zookeeper.

Change-Id: I14421d5363a32adfb157662d4795ba8890e2327f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/159
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>