Only try to lint the current commit
Since we're running buildkite on every commit now (by virtue of every
commit being its own CL), we don't need to run the commit_lint script on
every single commit between HEAD and master, we only need to run it on
HEAD (or, more specifically, the value of the buildkite commit
environment variable).

Also, this was broken because it was referencing the wrong branch.

Change-Id: I6eb9657f047ee3ef0a5a1908aba42b6a2d3e69b8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/141
Reviewed-by: Harley Klein <harley@readyset.io>
Tested-by: Buildkite CI