Abort the process if the important tasks fail
If any of the tokio tasks running the HTTP listener, the controller, or
the worker fails for any reason, whether it's a panic or a regular error
return, we don't want to get in an awkward state where the noria-server
process is still running, but it's not *functional* in any way because
these critical tasks aren't running. In this case, the best thing to do
is to abort the process entirely and let the general orchestration (read:
systemd) restart us automatically.

To this end, there's a new `abort_on_panic` helper function in
launchpad/src/futures which wraps a future in an adapter that causes it
to call `process::abort()` if it panics. We don't want *all* panics to
abort the process, so we're not using panic-as-abort or some kind of
custom panic handler, but this gives us a cheap opt-in on a per-future
basis for that behavior.

Change-Id: Iaa7d586636e386ef4ec48a248bd29a9b15a09bf3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/144
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>
Reviewed-by: Justin Miron <justin@readyset.io>