Split BackendBuilder into two stages
Split the construction of a noria_adapter Backend via a BackendBuilder
into two stages, one to configure clone-able flags and fields, and one
to attach a writer and reader, the latter passed via build(). This is
primarily a readability enhancement in constructing Backends for new
connectors, as we can configure all of the fields once out of the loop
and then just clone the entire builder into the future for each
constructor. Moving the reader and writer out was necessary because the
writer can't implement Clone due to eventually containing a Zookeeper
connection, which doesn't implement Clone and we don't have control
over.

Change-Id: Ie42d695cc9f4992b8ea4be68a06d68f9a624b617
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/147
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>