Add a flag to make adapter reads race
Add a new flag, `--race-reads`, to the adapter crate (and hence both
noria-mysql and noria-psql, now) which requires fallback and when set
will cause *all* reads to run on two parallel tasks on both mysql and
noria, returning the first successful result, or the mysql error if both
return an error. The goal here is to eventually use this mode to get
both coverage and benchmark results for the set of queries an end-user
runs in their application.

As of this commit, this is untested - and reasonably difficult to test
effectively as prepared statements don't go through this code path - so
it might not actually work yet. Once we have prepared statements going
through fallback this will be testable, and may change somewhat.

Change-Id: Ieef6995ab5c45b80b11b624a81946a1782e26bd7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/149
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>